Rename meta information to status
