--- hyrisecockpit/api/app/meta_information/interface.py
@@ -1,1 +1,0 @@
-"""Interfaces for meta information module."""
