+++ hyrisecockpit/api/app/routes.py
@@ -11,11 +11,11 @@ def register_routes(api: Api, app: Flask, root: str = "/api") -> None:
     from .monitor import register_routes as attach_monitor
     from .control import register_routes as attach_control
     from .sql import register_routes as attach_sql
-    from .meta_information import register_routes as attach_meta_information
+    from .status import register_routes as attach_status
 
     attach_workload(api, app, root)
     attach_database(api, app, root)
     attach_monitor(api, app, root)
     attach_control(api, app, root)
     attach_sql(api, app, root)
-    attach_meta_information(api, app, root)
+    attach_status(api, app, root)
