+++ hyrisecockpit/api/app/status/interface.py
@@ -1,0 +1,1 @@
+"""Interfaces for status information module."""
